Implement, tests more policies.

+++ app/Http/Controllers/FacilityController.php
@@ -23,10 +23,14 @@ class FacilityController extends Controller
     /**
      * Get all facilities.
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
+        $this->authorize('viewAny', [Facility::class]);
+
         $facilities = Facility::withTrashed()->get();
 
         return response(['facilities' => $facilities]);
@@ -38,11 +42,14 @@ class FacilityController extends Controller
      * @param \Illuminate\Http\Request $request
      *
      * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      *
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Facility::class]);
+
         $this->validate($request, [
             'name' => 'required|max:100',
             'description' => 'nullable|max:100',
@@ -74,10 +81,14 @@ class FacilityController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
+        $this->authorize('view', [Facility::class, $id]);
+
         $facility = Facility::withTrashed()->findOrFail($id);
 
         return response($facility);
@@ -90,11 +101,14 @@ class FacilityController extends Controller
      * @param string                   $id
      *
      * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      *
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', [Facility::class]);
+
         $this->validate($request, [
             'name' => 'sometimes|max:100',
             'description' => 'nullable|max:100',
@@ -126,12 +140,15 @@ class FacilityController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @throws \Exception
      *
      * @return \Illuminate\Http\Response
      */
     public function revoke($id)
     {
+        $this->authorize('soft-delete', [Facility::class]);
+
         $facility = Facility::findOrFail($id);
 
         $facility->delete();
@@ -146,10 +163,14 @@ class FacilityController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function restore($id)
     {
+        $this->authorize('restore', [Facility::class]);
+
         $facility = Facility::onlyTrashed()->findOrFail($id);
 
         $facility->restore();
@@ -164,10 +185,14 @@ class FacilityController extends Controller
      *
      * @param string $id
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        $this->authorize('force-delete', [Facility::class]);
+
         $facility = Facility::onlyTrashed()->findOrFail($id);
 
         $facility->forceDelete();
@@ -182,11 +207,14 @@ class FacilityController extends Controller
      * @param string                   $id
      *
      * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      *
      * @return \Illuminate\Http\Response
      */
     public function sync_modules(Request $request, $id)
     {
+        $this->authorize('assign-modules', [Module::class]);
+
         $facility = Facility::findOrFail($id);
 
         $this->validate($request, [
