+++ app/Http/Controllers/ModuleController.php
@@ -21,10 +21,14 @@ class ModuleController extends Controller
     /**
      * Get modules.
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
+        $this->authorize('viewAny', [Module::class]);
+
         $modules = Module::withTrashed()->get();
 
         return response(['modules' => $modules]);
@@ -36,11 +40,14 @@ class ModuleController extends Controller
      * @param \Illuminate\Http\Request $request
      *
      * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      *
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Module::class]);
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|max:20',
             'description' => 'sometimes|max:25',
@@ -75,10 +82,14 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function show($name)
     {
+        $this->authorize('view', [Module::class]);
+
         $module = Module::withTrashed()->findOrFail($name);
 
         return response($module);
@@ -91,11 +102,14 @@ class ModuleController extends Controller
      * @param string                   $name
      *
      * @throws ValidationException
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      *
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $name)
     {
+        $this->authorize('update', [Module::class]);
+
         $module = Module::findOrFail($name);
 
         $this->validate($request, [
@@ -115,12 +129,15 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @throws \Exception
      *
      * @return \Illuminate\Http\Response
      */
     public function revoke($name)
     {
+        $this->authorize('soft-delete', [Module::class]);
+
         $module = Module::findOrFail($name);
 
         $module->delete();
@@ -135,10 +152,14 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function restore($name)
     {
+        $this->authorize('restore', [Module::class]);
+
         $module = Module::onlyTrashed()->findOrFail($name);
 
         $module->restore();
@@ -153,10 +174,14 @@ class ModuleController extends Controller
      *
      * @param string $name
      *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy($name)
     {
+        $this->authorize('force-delete', [Module::class]);
+
         $module = Module::onlyTrashed()->findOrFail($name);
 
         $module->forceDelete();
